requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
openai==1.12.0
groq==0.4.2
requests==2.31.0
